tests/submission/PhilipTaylor/canvas/size.attributes.setAttribute.empty.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.setAttribute.empty</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.<a href="index.size.attributes.setAttribute.html">setAttribute</a>.empty</h1>
<p class="desc">Parsing of non-negative integers in setAttribute</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.nonnegativeinteger">size.nonnegativeinteger</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.size.error">size.error</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="size.attributes.setAttribute.empty.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Parsing of non-negative integers in setAttribute");
_addTest(function(canvas, ctx) {

canvas.setAttribute('width', '');
canvas.setAttribute('height', '');
_assertSame(canvas.width, 300, "canvas.width", "300");
_assertSame(canvas.height, 150, "canvas.height", "150");


});
</script>

<div id="log"></div>
</body></html>
